add ToSend partial class
